Event Page - Update searchParams to use asynchronous handling (Promise<NextServerSearchParams>), aligning with Next.js 15 changes.
