+++ apps/site/src/app/events/page.tsx
@@ -1,73 +1,41 @@
-import dynamic from 'next/dynamic'
 import Image from 'next/image'
-
-import { MagnifyingGlass } from '@phosphor-icons/react/dist/ssr'
 
 import type { AsyncNextServerSearchParams } from '@/types/searchParams'
 
 import { PATHS } from '@/constants/paths'
-import { CATEGORY_KEY, LOCATION_KEY } from '@/constants/searchParams'
 
 import { graphicsData } from '@/data/graphicsData'
 
 import { buildImageSizeProp } from '@/utils/buildImageSizeProp'
-import { getCategoryLabel } from '@/utils/categoryUtils'
 import { createMetadata } from '@/utils/createMetadata'
 import { extractSlugFromFilename } from '@/utils/fileUtils'
-import {
-  entryMatchesCategoryQuery,
-  entryMatchesLocationQuery,
-} from '@/utils/filterUtils'
 import { getFrontmatter } from '@/utils/getFrontmatter'
-import { normalizeQueryParam } from '@/utils/queryUtils'
 
 import { FeaturedPageFrontmatterSchema } from '@/schemas/FrontmatterSchema'
 
-import { useFilter } from '@/hooks/useFilter'
-import { useListboxOptions } from '@/hooks/useListboxOptions'
-import { usePagination } from '@/hooks/usePagination'
-import { useSearch } from '@/hooks/useSearch'
-import { useSort } from '@/hooks/useSort'
-
 import { Card } from '@/components/Card'
 import { CardGrid } from '@/components/CardGrid'
-import { CategoryFilter } from '@/components/CategoryFilter'
-import { FilterContainer } from '@/components/FilterContainer'
-import { LocationFilter } from '@/components/LocationFilter'
-import { NoSearchResultsMessage } from '@/components/NoSearchResultsMessage'
 import { PageHeader } from '@/components/PageHeader'
 import { PageLayout } from '@/components/PageLayout'
 import { PageSection } from '@/components/PageSection'
-import { Search } from '@/components/Search'
 import { StructuredDataScript } from '@/components/StructuredDataScript'
 
-import { EventSort } from './components/EventSort'
-import { DEFAULT_CTA_TEXT, FILTERS_CONFIG } from './constants/constants'
-import { eventsSortConfigs } from './constants/sortConfigs'
+import EventsContent from './components/EventsContent'
+import { DEFAULT_CTA_TEXT } from './constants/constants'
 import { getInvolvedData } from './data/getInvolvedData'
 import { generateStructuredData } from './utils/generateStructuredData'
 import { getEventData, getEventsData } from './utils/getEventData'
 import { getLocationListboxOptions } from './utils/getLocationFilterOptions'
 import { getMetaData } from './utils/getMetaData'
 
-const NoSSRPagination = dynamic(
-  () => import('@/components/Pagination').then((module) => module.Pagination),
-  { ssr: false },
-)
-
 type Props = {
   searchParams: AsyncNextServerSearchParams
 }
 
-const events = getEventsData()
-
-const { featuredEntry: featuredEventPath, seo } = getFrontmatter({
+const { seo, featuredEntry: featuredEventPath } = getFrontmatter({
   path: PATHS.EVENTS,
   zodParser: FeaturedPageFrontmatterSchema.parse,
 })
-
-const featuredEventSlug = extractSlugFromFilename(featuredEventPath)
-const featuredEvent = getEventData(featuredEventSlug)
 
 export const metadata = createMetadata({
   seo: {
@@ -78,7 +46,11 @@ export const metadata = createMetadata({
   overrideDefaultTitle: true,
 })
 
+const events = getEventsData()
+const featuredEventSlug = extractSlugFromFilename(featuredEventPath)
+const featuredEvent = getEventData(featuredEventSlug)
 const locationOptions = getLocationListboxOptions()
+const structuredData = generateStructuredData(seo)
 
 export default async function Events(props: Props) {
   const searchParams = await props.searchParams
@@ -87,46 +59,9 @@ export default async function Events(props: Props) {
     throw new Error('Featured event not found')
   }
 
-  const { searchQuery, searchResults } = useSearch({
-    searchParams,
-    entries: events,
-    searchBy: ['title', 'location'],
-  })
-
-  const { sortedResults } = useSort({
-    searchParams,
-    entries: searchResults,
-    configs: eventsSortConfigs,
-    defaultsTo: 'all-events',
-  })
-
-  const { filteredEntries: filteredEventsByLocation } = useFilter({
-    entries: sortedResults,
-    filterQuery: normalizeQueryParam(searchParams, LOCATION_KEY),
-    filterFn: entryMatchesLocationQuery,
-  })
-
-  const { filteredEntries } = useFilter({
-    entries: filteredEventsByLocation,
-    filterQuery: normalizeQueryParam(searchParams, CATEGORY_KEY),
-    filterFn: entryMatchesCategoryQuery,
-  })
-
-  const { optionsWithCount: categoryOptionsWithCount } = useListboxOptions({
-    collectionName: FILTERS_CONFIG.category.collectionName,
-    fieldName: FILTERS_CONFIG.category.fieldName,
-    defaultOption: FILTERS_CONFIG.category.defaultOption,
-    entries: filteredEventsByLocation,
-  })
-
-  const { currentPage, pageCount, paginatedResults } = usePagination({
-    searchParams,
-    entries: filteredEntries,
-  })
-
   return (
     <PageLayout>
-      <StructuredDataScript structuredData={generateStructuredData(seo)} />
+      <StructuredDataScript structuredData={structuredData} />
       <PageHeader
         isFeatured
         title={featuredEvent.title}
@@ -142,108 +77,16 @@ export default async function Events(props: Props) {
           objectFit: 'cover',
         }}
         cta={{
-          href: `${PATHS.EVENTS.path}/${featuredEventSlug}`,
+          href: `${PATHS.EVENTS.path}/${featuredEvent.slug}`,
           text: DEFAULT_CTA_TEXT,
         }}
       />
       <PageSection kicker="Events" title="Network Events">
-        <FilterContainer>
-          <FilterContainer.ResultsAndCategory
-            category={<CategoryFilter options={categoryOptionsWithCount} />}
-          />
-          <FilterContainer.MainWrapper>
-            <FilterContainer.DesktopFilters
-              searchComponent={<Search query={searchQuery} />}
-              sortComponent={<EventSort />}
-              filterComponents={[
-                <LocationFilter key="location" options={locationOptions} />,
-              ]}
-            />
-            <FilterContainer.MobileFiltersAndResults
-              searchComponent={<Search query={searchQuery} />}
-              sortComponent={<EventSort />}
-              filterComponents={[
-                <CategoryFilter
-                  key="category"
-                  options={categoryOptionsWithCount}
-                />,
-                <LocationFilter key="location" options={locationOptions} />,
-              ]}
-            />
-            <FilterContainer.ContentWrapper>
-              {filteredEntries.length === 0 ? (
-                <NoSearchResultsMessage />
-              ) : (
-                <>
-                  <CardGrid cols="smTwo">
-                    {paginatedResults.map((event, i) => {
-                      const {
-                        slug,
-                        title,
-                        image,
-                        category,
-                        description,
-                        externalLink,
-                        startDate,
-                        endDate,
-                        location,
-                      } = event
-
-                      const isFirstTwoImages = i < 2
-                      const shouldLinkToExternalEventsPage =
-                        !description && externalLink
-
-                      const tagLabel = getCategoryLabel({
-                        collectionName: 'event_entries',
-                        category,
-                      })
-
-                      return (
-                        <Card
-                          key={slug}
-                          title={title}
-                          borderColor="brand-400"
-                          textIsClamped={true}
-                          tags={[{ text: tagLabel }]}
-                          metaData={getMetaData({
-                            startDate,
-                            endDate,
-                            location: location.primary,
-                          })}
-                          image={{
-                            ...(image || graphicsData.imageFallback.data),
-                            alt: '',
-                            priority: isFirstTwoImages,
-                            objectFit: 'cover',
-                            sizes: buildImageSizeProp({
-                              startSize: '100vw',
-                              sm: '350px',
-                              md: '450px',
-                              lg: '360px',
-                            }),
-                          }}
-                          cta={{
-                            href:
-                              shouldLinkToExternalEventsPage ||
-                              `${PATHS.EVENTS.path}/${slug}`,
-                            text: DEFAULT_CTA_TEXT,
-                            icon: MagnifyingGlass,
-                          }}
-                        />
-                      )
-                    })}
-                  </CardGrid>
-                  <FilterContainer.PaginationWrapper>
-                    <NoSSRPagination
-                      pageCount={pageCount}
-                      currentPage={currentPage}
-                    />
-                  </FilterContainer.PaginationWrapper>
-                </>
-              )}
-            </FilterContainer.ContentWrapper>
-          </FilterContainer.MainWrapper>
-        </FilterContainer>
+        <EventsContent
+          searchParams={searchParams}
+          events={events}
+          locationOptions={locationOptions}
+        />
       </PageSection>
 
       <PageSection
