+++ package.json
@@ -6,11 +6,71 @@
   ],
   "packageManager": "npm@10.9.0",
   "scripts": {
+<<<<<<< HEAD
     "build": "turbo run build",
     "start": "turbo run start",
     "dev": "turbo run dev",
     "lint": "turbo run lint",
     "test": "turbo run test"
+=======
+    "predev": "node scripts/convertYamlToJson.js",
+    "dev": "next dev",
+    "prebuild": "node scripts/convertYamlToJson.js && node scripts/injectEncryptionAccessKeyInCMS.js",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint --fix",
+    "test:all": "cypress run --spec 'cypress/e2e/**/*.cy.ts'",
+    "test:cypress:percy": "percy exec -- cypress run --spec 'cypress/e2e/**/*.cy.ts'",
+    "test:cypress": "cypress run --spec 'cypress/e2e/**/*.cy.ts'",
+    "generate:page": "node scripts/generatePageTemplates.js",
+    "prepare": "husky"
+  },
+  "dependencies": {
+    "@date-fns/utc": "^2.1.0",
+    "@headlessui/react": "^2.2.0",
+    "@headlessui/tailwindcss": "^0.2.1",
+    "@hookform/resolvers": "^3.9.1",
+    "@octokit/rest": "^21.0.1",
+    "@phosphor-icons/react": "^2.1.7",
+    "@radix-ui/react-popover": "^1.1.4",
+    "@sentry/nextjs": "^8.24.0",
+    "@tailwindcss/forms": "^0.5.9",
+    "@tailwindcss/typography": "^0.5.15",
+    "@tanstack/react-table": "^8.20.6",
+    "@vercel/speed-insights": "^1.1.0",
+    "airtable": "^0.12.2",
+    "autoprefixer": "^10.4.20",
+    "camelcase-keys": "^9.1.3",
+    "clsx": "^2.1.1",
+    "crypto-js": "^4.2.0",
+    "date-fns": "^4.1.0",
+    "decap-cms-app": "^3.4.0",
+    "frontmatter-markdown-loader": "^3.7.0",
+    "gray-matter": "^4.0.3",
+    "joi": "^17.13.3",
+    "js-yaml": "^4.1.0",
+    "netlify-identity-widget": "^1.9.2",
+    "next": "15.1.2",
+    "next-plausible": "^3.12.4",
+    "nuqs": "^2.2.3",
+    "pluralize": "^8.0.0",
+    "pretty-bytes": "^6.1.1",
+    "ramda": "^0.30.1",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "react-hook-form": "^7.54.2",
+    "react-lite-youtube-embed": "^2.4.0",
+    "react-markdown": "^9.0.1",
+    "rehype-raw": "^7.0.0",
+    "remark-gfm": "^4.0.0",
+    "remove-markdown": "^0.6.0",
+    "schema-dts": "^1.1.2",
+    "slugify": "^1.6.6",
+    "swr": "^2.3.0",
+    "tailwindcss": "^3.4.17",
+    "usehooks-ts": "^3.1.0",
+    "zod": "^3.24.1"
+>>>>>>> 95f381e15 (Event Page - Update searchParams to use asynchronous handling (Promise<NextServerSearchParams>), aligning with Next.js 15 changes.)
   },
   "devDependencies": {
     "turbo": "^2.3.3"
